ask/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coverage==6.1.1
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-filter==21.1
django-js-asset==1.2.2
django-seed==0.3.1
django-taggit==1.5.1
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==9.8.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
urllib3==1.26.7
